+++ src/main/java/top/offsetmonkey538/offsetend/datagen/DataGeneration.java
@@ -3,6 +3,7 @@ package top.offsetmonkey538.offsetend.datagen;
 import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
 import top.offsetmonkey538.offsetend.datagen.language.ModEnglishLanguageProvider;
+import top.offsetmonkey538.offsetend.datagen.tag.ModBlockTagProvider;
 
 public class DataGeneration implements DataGeneratorEntrypoint {
 
@@ -11,9 +12,13 @@ public class DataGeneration implements DataGeneratorEntrypoint {
         FabricDataGenerator.Pack pack = fabricDataGenerator.createPack();
 
         pack.addProvider(ModModelProvider::new);
-        pack.addProvider(ModBlockTagProvider::new);
-        pack.addProvider(ModEnglishLanguageProvider::new);
         pack.addProvider(ModBlockLootTableProvider::new);
         pack.addProvider(ModRecipeProvider::new);
+
+        // Tags
+        pack.addProvider(ModBlockTagProvider::new);
+
+        // Language
+        pack.addProvider(ModEnglishLanguageProvider::new);
     }
 }
